requirements.txt
fastapi
pydantic
llama-cpp-python
python-dotenv
uvicorn
